Fixed forced login

+++ app/addresses/mutations/deleteAddress.ts
@@ -1,5 +1,5 @@
 import getCurrentUser from "app/users/queries/getCurrentUser"
-import { resolver } from "blitz"
+import { AuthenticationError, resolver } from "blitz"
 import db from "db"
 
 import { DeleteAddress } from "../validations"
@@ -9,6 +9,7 @@ export default resolver.pipe(
   resolver.authorize(),
   async ({ id }, ctx) => {
     const user = await getCurrentUser(null, ctx)
+    if (!user) throw new AuthenticationError()
 
     const address = await db.address.deleteMany({ where: { id, userId: user.id } })
